# openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "properties": {
              "code": {
                "type": "string"
              },
              "details": {
                "nullable": true,
                "type": "object"
              },
              "message": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message"
            ],
            "type": "object"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "IngredientGroupInput": {
        "properties": {
          "ingredients": {
            "items": {
              "$ref": "#/components/schemas/IngredientInput"
            },
            "maxItems": 50,
            "minItems": 1,
            "title": "Ingredients",
            "type": "array"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 40,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "required": [
          "ingredients"
        ],
        "title": "IngredientGroupInput",
        "type": "object"
      },
      "IngredientGroupOutput": {
        "properties": {
          "ingredients": {
            "items": {
              "$ref": "#/components/schemas/IngredientOutput"
            },
            "title": "Ingredients",
            "type": "array"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "required": [
          "name",
          "ingredients"
        ],
        "title": "IngredientGroupOutput",
        "type": "object"
      },
      "IngredientInput": {
        "properties": {
          "name": {
            "maxLength": 60,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "quantity": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,7}|(?=[\\d.]{1,11}0*$)\\d{0,7}\\.\\d{0,3}0*$)",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quantity"
          },
          "refRecipeId": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Refrecipeid"
          },
          "unit": {
            "anyOf": [
              {
                "maxLength": 20,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          }
        },
        "required": [
          "name"
        ],
        "title": "IngredientInput",
        "type": "object"
      },
      "IngredientOutput": {
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "placement": {
            "title": "Placement",
            "type": "string"
          },
          "quantity": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quantity"
          },
          "refRecipe": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RefRecipe"
              },
              {
                "type": "null"
              }
            ]
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          }
        },
        "required": [
          "name",
          "quantity",
          "unit",
          "placement",
          "refRecipe"
        ],
        "title": "IngredientOutput",
        "type": "object"
      },
...
      "RecipeAuthor": {
        "properties": {
          "avatarUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatarurl"
          },
          "displayName": {
            "title": "Displayname",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "displayName"
        ],
        "title": "RecipeAuthor",
        "type": "object"
      },
      "RecipeFeedItem": {
        "description": "ホームフィードのレシピカード 1 枚分（features/home-feed.md §5）。\n\n自分のレシピ一覧（`RecipeSummary`）と違い、他人のレシピも並ぶので\n投稿者情報（`author`）を含める。`is_public` は公開レシピしか出さないため\n持たせない。`favorite_count` は `recipes.favorite_count`（カウント列\nキャッシュ）をそのまま返す（Phase 6 でお気に入り機能が入るまでは 0）。",
        "properties": {
          "author": {
            "$ref": "#/components/schemas/RecipeAuthor"
          },
          "favoriteCount": {
            "title": "Favoritecount",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "thumbnailUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnailurl"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "thumbnailUrl",
          "author",
          "favoriteCount"
        ],
        "title": "RecipeFeedItem",
        "type": "object"
      },
      "RecipeFeedResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RecipeFeedItem"
            },
            "title": "Items",
            "type": "array"
          },
          "nextCursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nextcursor"
          }
        },
        "required": [
          "items",
          "nextCursor"
        ],
        "title": "RecipeFeedResponse",
        "type": "object"
      },
      "RecipeListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RecipeSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "nextCursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nextcursor"
          }
        },
        "required": [
          "items",
          "nextCursor"
        ],
        "title": "RecipeListResponse",
        "type": "object"
      },
      "RecipeResponse": {
        "properties": {
          "author": {
            "$ref": "#/components/schemas/RecipeAuthor"
          },
          "commentCount": {
            "title": "Commentcount",
            "type": "integer"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "favoriteCount": {
            "title": "Favoritecount",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "ingredientGroups": {
            "items": {
              "$ref": "#/components/schemas/IngredientGroupOutput"
            },
            "title": "Ingredientgroups",
            "type": "array"
          },
          "isFavorited": {
            "title": "Isfavorited",
            "type": "boolean"
          },
          "isPublic": {
            "title": "Ispublic",
            "type": "boolean"
          },
          "servings": {
            "title": "Servings",
            "type": "integer"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/StepOutput"
            },
            "title": "Steps",
            "type": "array"
          },
          "thumbnailKey": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnailkey"
          },
          "thumbnailUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnailurl"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "title": "Updatedat",
            "type": "string"
          }
        },
        "required": [
          "id",
          "author",
          "title",
          "description",
          "servings",
          "isPublic",
          "thumbnailUrl",
          "thumbnailKey",
          "isFavorited",
          "favoriteCount",
          "commentCount",
          "ingredientGroups",
          "steps",
          "createdAt",
          "updatedAt"
        ],
        "title": "RecipeResponse",
        "type": "object"
      },
      "RecipeSummary": {
        "description": "自分のレシピ一覧のカード 1 枚分（features/recipe.md §2「自分のレシピ一覧」）。",
        "properties": {
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "isPublic": {
            "title": "Ispublic",
            "type": "boolean"
          },
          "thumbnailUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnailurl"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "thumbnailUrl",
          "isPublic",
          "createdAt"
        ],
        "title": "RecipeSummary",
        "type": "object"
      },
      "RecipeWriteRequest": {
        "description": "`POST /recipes` と `PUT /recipes/{id}` 共通のリクエスト body。\n\n`PUT` の「`thumbnailKey` 省略 = サムネイル変更なし」は、ルーター側で\n`model_fields_set` に `\"thumbnail_key\"` が含まれるかで判定する\n（features/recipe.md §5「省略 = 変更なし / null = 削除 / 既存キー再送 = 維持 /\n新キー = 差し替え」）。",
        "properties": {
          "description": {
            "default": "",
            "maxLength": 2000,
            "title": "Description",
            "type": "string"
          },
          "ingredientGroups": {
            "items": {
              "$ref": "#/components/schemas/IngredientGroupInput"
            },
            "maxItems": 20,
            "minItems": 1,
            "title": "Ingredientgroups",
            "type": "array"
          },
          "isPublic": {
            "default": false,
            "title": "Ispublic",
            "type": "boolean"
          },
          "servings": {
            "maximum": 99.0,
            "minimum": 1.0,
            "title": "Servings",
            "type": "integer"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/StepInput"
            },
            "maxItems": 100,
            "minItems": 1,
            "title": "Steps",
            "type": "array"
          },
          "thumbnailKey": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnailkey"
          },
          "title": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "servings",
          "ingredientGroups",
          "steps"
        ],
        "title": "RecipeWriteRequest",
        "type": "object"
      },
      "RefRecipe": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title"
        ],
        "title": "RefRecipe",
        "type": "object"
      },
...
      "StepInput": {
        "properties": {
          "body": {
            "maxLength": 1000,
            "title": "Body",
            "type": "string"
          },
          "imageKey": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Imagekey"
          }
        },
        "required": [
          "body"
        ],
        "title": "StepInput",
        "type": "object"
      },
      "StepOutput": {
        "properties": {
          "body": {
            "title": "Body",
            "type": "string"
          },
          "imageKey": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Imagekey"
          },
          "imageUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Imageurl"
          }
        },
        "required": [
          "body",
          "imageUrl",
          "imageKey"
        ],
        "title": "StepOutput",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Recipi API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/recipes": {
      "get": {
        "description": "ホームフィード / 検索（features/home-feed.md・search.md）。\n\nMVP で受け付ける `feed` は `all` のみ。`following` / `followers` / `favorites`\nは Phase 5・6 で有効化するので、それらを含む `all` 以外の値は 400 にする\n（home-feed.md §6「不正値は 400」）。",
        "operationId": "list_feed_api_v1_recipes_get",
        "parameters": [
          {
            "in": "query",
            "name": "feed",
            "required": false,
            "schema": {
              "default": "all",
              "title": "Feed",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeFeedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "リクエストの内容が不正です"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Feed",
        "tags": [
          "recipes"
        ]
      },
      "post": {
        "operationId": "create_recipe_api_v1_recipes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeWriteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "リクエストの内容が不正です"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Recipe",
        "tags": [
          "recipes"
        ]
      }
    },
    "/api/v1/recipes/{recipe_id}": {
      "delete": {
        "operationId": "delete_recipe_api_v1_recipes__recipe_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "recipe_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Recipe Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "リクエストの内容が不正です"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Recipe",
        "tags": [
          "recipes"
        ]
      },
      "get": {
        "operationId": "get_recipe_api_v1_recipes__recipe_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "recipe_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Recipe Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "リクエストの内容が不正です"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {},
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Recipe",
        "tags": [
          "recipes"
        ]
      },
      "put": {
        "operationId": "update_recipe_api_v1_recipes__recipe_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "recipe_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Recipe Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeWriteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "リクエストの内容が不正です"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Recipe",
        "tags": [
          "recipes"
        ]
      }
    },
    "/api/v1/recipes/{recipe_id}/view": {
      "post": {
        "description": "レシピ詳細を開いた記録を残す（features/view-history.md §3）。\n\nクライアントは `GET /recipes/{id}` 成功後に非同期で 1 回だけ呼ぶ。\n見えないレシピは 404、成功は 204（body なし）。書き込みなので commit を明示する。",
        "operationId": "record_recipe_view_api_v1_recipes__recipe_id__view_post",
        "parameters": [
          {
            "in": "path",
            "name": "recipe_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Recipe Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "リクエストの内容が不正です"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Record Recipe View",
        "tags": [
          "recipes"
        ]
      }
    },
...
    "/api/v1/users/me/recipes": {
      "get": {
        "operationId": "list_my_recipes_api_v1_users_me_recipes_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "リクエストの内容が不正です"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List My Recipes",
        "tags": [
          "recipes"
        ]
      }
    },
...
  }
}